+++ src/lib.rs
@@ -57,5 +57,3 @@ pub mod utils;
 
 mod cache;
 mod testing;
-
-mod jsonrpc_error;
