+++ src/utils.rs
@@ -1,9 +1,11 @@
 use std::collections::HashMap;
 use std::convert::TryInto;
+use std::fmt;
 use std::pin::Pin;
 
 use chrono::{DateTime, Utc};
 use futures::Future;
+use jsonrpc_core::{Error, ErrorCode};
 use multivm::interface::{ExecutionResult, VmExecutionResultAndLogs, VmInterface};
 use multivm::vm_latest::HistoryDisabled;
 use multivm::vm_latest::Vm;
@@ -225,6 +227,44 @@ pub fn utc_datetime_from_epoch_ms(millis: u64) -> DateTime<Utc> {
     let nanos = (millis % 1000) * 1_000_000;
     // expect() is ok- nanos can't be >2M
     DateTime::<Utc>::from_timestamp(secs as i64, nanos as u32).expect("valid timestamp")
+}
+
+pub fn into_jsrpc_error(err: Web3Error) -> Error {
+    Error {
+        code: match err {
+            Web3Error::InternalError | Web3Error::NotImplemented => ErrorCode::InternalError,
+            Web3Error::NoBlock
+            | Web3Error::PrunedBlock(_)
+            | Web3Error::PrunedL1Batch(_)
+            | Web3Error::NoSuchFunction
+            | Web3Error::RLPError(_)
+            | Web3Error::InvalidTransactionData(_)
+            | Web3Error::TooManyTopics
+            | Web3Error::FilterNotFound
+            | Web3Error::InvalidFeeParams(_)
+            | Web3Error::LogsLimitExceeded(_, _, _)
+            | Web3Error::InvalidFilterBlockHash => ErrorCode::InvalidParams,
+            Web3Error::SubmitTransactionError(_, _) | Web3Error::SerializationError(_) => 3.into(),
+            Web3Error::PubSubTimeout => 4.into(),
+            Web3Error::RequestTimeout => 5.into(),
+            Web3Error::TreeApiUnavailable => 6.into(),
+        },
+        message: match err {
+            Web3Error::SubmitTransactionError(_, _) => err.to_string(),
+            _ => err.to_string(),
+        },
+        data: match err {
+            Web3Error::SubmitTransactionError(_, data) => {
+                Some(format!("0x{}", hex::encode(data)).into())
+            }
+            _ => None,
+        },
+    }
+}
+
+pub fn internal_error(method_name: &'static str, error: impl fmt::Display) -> Web3Error {
+    tracing::error!("Internal error in method {method_name}: {error}");
+    Web3Error::InternalError
 }
 
 #[cfg(test)]
